[FEAT] make shareprofile pop up spareate componet
